refactor(services): split user services tests in multiple files
